Attempt to split the Board module into Board 'role' modules
@maikon
@jsuchy
@felipesere
I thought this might be a good idea, but maybe it's not.. let me know
